+++ lore-revision/src/store/remote.rs
@@ -284,8 +284,6 @@ impl store::ImmutableStore for RemoteImmutableStore {
         None
     }
 
-    async fn compact_stop(self: Arc<Self>) {}
-
     async fn verify(self: Arc<Self>, _heal: bool) -> Result<(), StoreError> {
         Ok(())
     }