+++ lore-storage/src/defragment.rs
@@ -2253,10 +2253,6 @@ mod tests {
                 self.inner.clone().compact_resume_at().await
             }
 
-            async fn compact_stop(self: Arc<Self>) {
-                self.inner.clone().compact_stop().await;
-            }
-
             fn max_query_batch(&self) -> Option<usize> {
                 None
             }